Extract default story out into JS file
